# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Partzona Partner API",
        "description": "1688.com məhsul məlumatlarına çıxış üçün Partner API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "Partner API v1"
        }
    ],
    "paths": {
...
        "/orders/{orderId}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Sifariş detalları",
                "description": "Konkret sifarişin tam məlumatlarını qaytarır: status, məhsullar, qiymətlər, tracking.",
                "operationId": "getOrderDetail",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/lang"
                    },
                    {
                        "name": "orderId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sifariş detalları"
                    },
                    "401": {
                        "description": "Token etibarsızdır"
                    },
                    "429": {
                        "description": "Rate limit aşılıb"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Orders"
                ],
                "summary": "Sifarişi ləğv et",
                "description": "Ödənilməmiş sifarişi ləğv edir.",
                "operationId": "cancelOrder",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/lang"
                    },
                    {
                        "name": "orderId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "reason",
                        "in": "query",
                        "description": "Ləğv səbəbi",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sifariş ləğv edildi"
                    },
                    "401": {
                        "description": "Token etibarsızdır"
                    },
                    "429": {
                        "description": "Rate limit aşılıb"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/orders/{orderId}/confirm": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Malı qəbul et",
                "description": "Sifarişi 'alındı' olaraq işarələyir. Geri qaytarıla bilməz!",
                "operationId": "confirmReceipt",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/lang"
                    },
                    {
                        "name": "orderId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mal qəbul edildi"
                    },
                    "401": {
                        "description": "Token etibarsızdır"
                    },
                    "429": {
                        "description": "Rate limit aşılıb"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/orders/{orderId}/refund": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Refund yarat",
                "description": "Ödənilmiş sifariş üçün geri qaytarma müraciəti yaradır.",
                "operationId": "createRefund",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/lang"
                    },
                    {
                        "name": "orderId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "entry_ids"
                                ],
                                "properties": {
                                    "entry_ids": {
                                        "description": "Geri qaytarılacaq entry ID-lər",
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": {
                                        "description": "refund=yalnız pul, returnRefund=mal+pul",
                                        "type": "string",
                                        "enum": [
                                            "refund",
                                            "returnRefund"
                                        ]
                                    },
                                    "reason": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "amount": {
                                        "type": "number",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Refund müraciəti yaradıldı"
                    },
                    "401": {
                        "description": "Token etibarsızdır"
                    },
                    "422": {
                        "description": "Validasiya xətası"
                    },
                    "429": {
                        "description": "Rate limit aşılıb"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "parameters": {
            "lang": {
                "name": "lang",
                "in": "query",
                "description": "Cavab dilini təyin edir (az, ru, en). Həmçinin Accept-Language header ilə göndərilə bilər.",
                "schema": {
                    "type": "string",
                    "default": "en",
                    "enum": [
                        "az",
                        "ru",
                        "en"
                    ]
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Partner API token. Authorization: Bearer pz_live_xxxxx",
                "scheme": "bearer"
            }
        }
    },
...
}